update: Move not found to component
